Update Authorization Server Profile.md
Text suggestion for resolving issue #7.

The inconsistency w.r.t. the sub claim results from extending the iGov profile. The iGov profile supports different use cases, which makes an optional sub claim logical. The NL-gov profile currently only supports a use case for which the presence of a sub claim will be required. The extension was too minimal, leading to an inconsistency. This proposed change updates the text on the sub claim, resolving the inconsistency.